Add test on all enedis API routes
